修正一个非数据类型的报错

+++ MLP/MLP_Iris.py
@@ -75,7 +75,8 @@ if __name__ == '__main__':
     data.iloc[np.where(data['Species'] == 'Iris-setosa')[0], -1] = 0
     data.iloc[np.where(data['Species'] == 'Iris-versicolor')[0], -1] = 1
     data.iloc[np.where(data['Species'] == 'Iris-virginica')[0], -1] = 2
-
+    # 将Species列的数据设置类型为int
+    data['Species'] = data['Species'].astype(int)
     # 数据集特征
     features = data[['SepalLengthCm',
                      'SepalWidthCm',
